huy upload code moi nhat

+++ admin/pages/administrator.php
@@ -6,7 +6,7 @@
                 <div class="page-header card">
                     <div class="card-block">
                         <h4 class="m-b-10">Quản trị viên</h4>
-                        <?php if ($_GET['control'] != 'addadmin') { ?>
+                        <?php if ($_GET['control'] != 'addadmin' && $_SESSION['login_admin']['id_role']==1) { ?>
                             <a href="index.php?control=addadmin"><button class="btn btn-primary mb-2">Thêm mới quản trị viên</button></a>
                         <?php } ?>
                         <!-- <p class="text-muted m-b-10">lorem ipsum dolor sit amet, consectetur adipisicing elit</p> -->
@@ -193,8 +193,14 @@
                                                 <th>#</th>
                                                 <th>Tài khoản</th>
                                                 <th>Họ tên</th>
-                                                <th>Sửa</th>
-                                                <th>Xóa</th>
+                                                <th>Tên cửa hàng</th>
+                                                <th>Email</th>
+                                                <th>Sđt</th>
+                                                <?php if($_SESSION['login_admin']['id_role']==1){ ?>
+                                                    <th>Sửa</th>
+                                                    <th>Xóa</th>     
+                                                <?php }?>
+                                                
                                             </tr>
                                         </thead>
                                         <tbody>
@@ -205,11 +211,15 @@
                                                     <td><?php echo $row['id_admin'] ?></td>
                                                     <td><?php echo $row['username'] ?></td>
                                                     <td><?php echo $row['fullname'] == "" ? "Chưa cập nhập" : $row['fullname'] ?></td>
-                                                    <td><a href="index.php?control=editadmin&id_admin=<?php echo $row['id_admin']?>" class="btn btn-warning">Sửa</a></td>
-                                                    <td><a href="index.php?control=deleteadmin&id_admin=<?php echo $row['id_admin']?>" class="btn btn-danger">Xóa</a></td>
+                                                    <td><?php echo $row['name_brand'] == "" ? "Chưa cập nhập" : $row['name_brand'] ?></td>
+                                                    <td><?php echo $row['email'] == "" ? "Chưa cập nhập" : $row['email'] ?></td>
+                                                    <td><?php echo $row['phone'] == "" ? "Chưa cập nhập" : $row['phone'] ?></td>
+                                                    <?php if($_SESSION['login_admin']['id_role']==1){ ?>
+                                                        <td><a href="index.php?control=editadmin&id_admin=<?php echo $row['id_admin']?>" class="btn btn-warning">Sửa</a></td>
+                                                        <td><a href="index.php?control=deleteadmin&id_admin=<?php echo $row['id_admin']?>" class="btn btn-danger">Xóa</a></td>
+                                                    <?php } ?>
                                                 </tr>
                                             <?php } ?>
-
                                         </tbody>
                                     </table>
                                 </div>
